Add ability to disable opflex-agent features
The opflex-agent is dependent on the vSwitch from the underling
distribution. Some distributions ship older versions of the vSwitch
which aren't capable of supporting new features in the opflex-agent,
and therefore those features have to be disabled. This patch adds
a new configuration item for this, which provides a string with a
comma-delimited list inside for the features that should be
disabled in the deployment.

The initial feature disabled is ERSPAN.