Fix OSP10 => OSP13 FFU
During fast-forward upgrades (FFUs), there is a transition from
using the systemd opflex-agent running on the hypervisor, and the
containerized opflex-agent. We start the opflex-agent while we still
have the systemd opflex-agent running. This causes the containerized
opflex-agent to crash, as it's trying to access some of the same
resources (e.g. sockets) as the systemd one, which causes supervisord
in the container to stop the opflex-agent process.

This patch does a restart of the opflex-agent container after bringing
down the systemd opflex-agent, so that the containerized one can then
come back up cleanly.