Support -f argument in build script
The -f argument was supported in OSP13, because the docker
registry was available on the host. In OSP16, that was removed,
so the URL supported for the local ACI repo had to be changed
from a http method to a local file. This adds support back
for using http, provided that the -s options is used.