Fix OSP10 to OSP13 FFU
When doing an OSP10 to OSP13 Fast-Forward Upgrade (FFU), the
steps remove the old "vlan<vlan ID" interface at the beginning
of an upgrade_task. The interface isn't restored until much later.
This fixes that, by ensuring that the ifup on the new interface
that replaces it happens earller on in the process, minimizing
down time.