Change container start order for deployment
During deployment (initial or upgrade), we want to start the
neutron-opflex-agent container after the opflex-agent container.
Commit 3392a1f52a02ddea3c57c4ffb7c63b5182d5aaed to the puppet
repo train branch (https://github.com/noironetworks/ciscoaci-puppet)
ensures that the opflex-agent container is restarted when the
neutron-opflex-agent container is (re-)started, so that the
two agents don't conflict with the vSwitch, and flow-mods get
programmed correctly.