Change container start order for deployment
During deployment (initial or upgrade), we want to start the
neutron-opflex-agent container after the opflex-agent container.
Commit 338e2eb72db3d160d189e98f86f2fc7e6b49d4f1 to the puppet
repo queens branch (https://github.com/noironetworks/ciscoaci-puppet)
ensures that the opflex-agent container is restarted when the
neutron-opflex-agent container is (re-)started, so that the
two agents don't conflict with the vSwitch, and flow-mods get
programmed correctly.