Fix creation of new directory
The restarts subdirectory wasn't getting created on minor updates.
This is because the condition only checks for the parent directory,
which would have been created from a previous release. This patch
moves the new directory to its own block with its own precondition,
in order to ensure that the upgrade happens.