Fix keystone notification pool
Commit 91b83fe introduced the ability to configure the pool used
for listening for notifications from Keystone. The default value
used was None, but was incorrectly treated as a "None". The correct
default should have been an empty string (''). This patch fixes
the default setting.

Note that rabbitmq doesn't seem to automatically delete any old
queues if they no longer have listeners, so they will have to be
deleted manually.